+++ index.ts
@@ -2,6 +2,7 @@ import type { CreateClient, CreateOptions } from './lib/types';
 import { apiUrls } from './lib/utils/apis';
 import { createEvents } from './lib/events';
 import { createSearch } from './lib/search';
+import { createSharing } from './lib/sharing';
 import { createSights } from './lib/sights';
 
 const _ = require('underscore');
@@ -111,6 +112,7 @@ export const createClient: CreateClient = function (clientOptions) {
     request: require('./lib/request/').create(options),
     search: createSearch(options),
     server: require('./lib/server/').create(options),
+    sharing: createSharing(options),
     sheets: require('./lib/sheets/').create(options),
     sights: createSights(options),
     templates: require('./lib/templates/').create(options),
@@ -129,3 +131,4 @@ export const smartSheetURIs = {
 
 export { CreateClient, CreateClientOptions, SmartsheetClient } from './lib/types';
 export * from './lib/events/types';
+export * from './lib/sharing';