+++ lib/sights/types.ts
@@ -284,10 +284,9 @@ export interface SightsApi {
   moveSight: MoveSight;
   getSightPublshStatus: GetSightPublishStatus;
   setSightPublishStatus: SetSightPublishStatus;
-  // TODO -> the sharing method will be exposed as part of the sharing module work
-  getShare: any;
-  listShares: any;
-  share: any;
-  deleteShare: any;
-  updateShare: any;
+  // Deprecated sharing methods - will be replaced by the sharing module
+  getSightShares?: any;
+  shareSight?: any;
+  deleteShare?: any;
+  updateShare?: any;
 }