+++ lib/types/SmartsheetClient.ts
@@ -1,5 +1,6 @@
 import type { EventsApi } from '../events/types';
 import type { SearchApi } from '../search/types';
+import type { SharingApi } from '../sharing';
 import type { SightsApi } from '../sights/types';
 
 export interface SmartsheetClient {
@@ -15,6 +16,7 @@ export interface SmartsheetClient {
   request: any;
   search: SearchApi;
   server: any;
+  sharing: SharingApi;
   sheets: any;
   sights: SightsApi;
   templates: any;